# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker API",
    "version": "1.0.0",
    "description": "API documentation for the Workout Tracker application"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/exercises": {
      "get": {
        "summary": "Get all exercises",
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "description": "Limit the number of exercises returned",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of exercises",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No exercises found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/exercises/{exercise_id}": {
      "get": {
        "summary": "Get exercise by ID",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "description": "ID of the exercise to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of exercise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayExercise"
                }
              }
            }
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/progress": {
      "post": {
        "summary": "Create a new progress entry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgress"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Progress created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayProgress"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all progress entries for the current user",
        "responses": {
          "200": {
            "description": "Successful retrieval of progress entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayProgress"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No progress found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/progress/{progress_id}": {
      "get": {
        "summary": "Get progress entry by ID",
        "parameters": [
          {
            "name": "progress_id",
            "in": "path",
            "description": "ID of the progress entry to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of progress entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayProgress"
                }
              }
            }
          },
          "404": {
            "description": "Progress entry not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update a progress entry",
        "parameters": [
          {
            "name": "progress_id",
            "in": "path",
            "description": "ID of the progress entry to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgress"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progress updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayProgress"
                }
              }
            }
          },
          "404": {
            "description": "Progress entry not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a progress entry",
        "parameters": [
          {
            "name": "progress_id",
            "in": "path",
            "description": "ID of the progress entry to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Progress deleted successfully"
          },
          "404": {
            "description": "Progress entry not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/workout_logs/{workout_log_id}/exercises": {
      "post": {
        "summary": "Add an exercise to a workout log",
        "parameters": [
          {
            "name": "workout_log_id",
            "in": "path",
            "description": "ID of the workout log to add exercise to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AddExerciseToWorkoutLog"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout log exercise created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutLogExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workout log not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get exercises logged for a specific workout log",
        "parameters": [
          {
            "name": "workout_log_id",
            "in": "path",
            "description": "ID of the workout log to retrieve exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of workout log exercises",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutLogExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No exercises logged for this workout"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/workouts/days/{day_id}/exercises": {
      "get": {
        "summary": "Get exercises in a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to retrieve exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of exercises",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Day not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Add an exercise to a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to add exercise to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseToWorkout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                }
              }
            }
          },
          "404": {
            "description": "Day or Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update exercises in a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to update exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateExerciseInWorkout"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercises updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Day or Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/days/{day_id}/exercises/{exercise_id}": {
      "delete": {
        "summary": "Delete an exercise from a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "exercise_id",
            "in": "path",
            "description": "ID of the exercise to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exercise deleted successfully"
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DisplayExercise": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "muscle_group": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "difficulty_level": {
            "type": "string",
            "nullable": true
          },
          "equipment_needed": {
            "type": "boolean"
          },
          "equipment_details": {
            "type": "string",
            "nullable": true
          },
          "calories_burnt_per_minute": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["exercise_id", "name", "equipment_needed", "created_at", "updated_at"]
      },
...
      "CreateProgress": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "weight": {
            "type": "number",
            "format": "float"
          },
          "bmi": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "body_fat_percentage": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "muscle_mass": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["date", "weight"]
      },
      "DisplayProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "weight": {
            "type": "number",
            "format": "float"
          },
          "bmi": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "body_fat_percentage": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "muscle_mass": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["id", "user_id", "date", "weight"]
      },
...
      "AddExerciseToWorkoutLog": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "sets_completed": {
            "type": "integer"
          },
          "reps_completed": {
            "type": "integer"
          },
          "weight_used": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "required": ["exercise_id", "sets_completed", "reps_completed"]
      },
      "DisplayWorkoutLogExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "workout_log_id": {
            "type": "integer"
          },
          "exercise_id": {
            "type": "integer"
          },
          "sets_completed": {
            "type": "integer"
          },
          "reps_completed": {
            "type": "integer"
          },
          "weight_used": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "required": ["id", "workout_log_id", "exercise_id", "sets_completed", "reps_completed"]
      },
...
      "AddExerciseToWorkout": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "workout_plan_day_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer"
          },
          "reps": {
            "type": "integer"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": ["exercise_id", "workout_plan_day_id", "sets", "reps", "order"]
      },
      "UpdateExerciseInWorkout": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer",
            "nullable": true
          },
          "reps": {
            "type": "integer",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["exercise_id"]
      },
      "DisplayWorkoutPlanExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "workout_plan_day_id": {
            "type": "integer"
          },
          "exercise_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer"
          },
          "reps": {
            "type": "integer"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": ["id", "workout_plan_day_id", "exercise_id", "sets", "reps", "order"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}